backend/reprocess-images/requirements.txt
Pillow==10.1.0
boto3==1.34.18
